[chore] move gemm_sp to experimental
